Show logged-in user email in sidebar footer
- Extract AppSidebar to its own client component file
- Layout server component fetches /api/users/me via strapiAdmin helper
- Add plugin::users-permissions.user.me to DonationAdmin bootstrap permissions
- Single auth cookie; no extra email cookie needed
